feature: add db cleaner; issue #566

+++ bcs-services/bcs-storage/app/options/options.go
@@ -42,9 +42,9 @@ type StorageOptions struct {
 
 	DBConfig     string `json:"database_config_file" value:"storage-database.conf" usage:"Config file for database."`
 	EventMaxTime int64  `json:"event_max_day" value:"15" usage:"Max day for holding events data."`
-	EventMaxCap  int    `json:"event_max_cap" value:"10000" usage:"Max num for holding events data of each cluster."`
+	EventMaxCap  int64  `json:"event_max_cap" value:"10000" usage:"Max num for holding events data of each cluster."`
 	AlarmMaxTime int64  `json:"alarm_max_day" value:"15" usage:"Max day for holding alarms data."`
-	AlarmMaxCap  int    `json:"alarm_max_cap" value:"10000" usage:"Max num for holding alarms data of each cluster."`
+	AlarmMaxCap  int64  `json:"alarm_max_cap" value:"10000" usage:"Max num for holding alarms data of each cluster."`
 	QueryMaxNum  int64  `json:"query_max_num" value:"100" usage:"Max num query to same url one time."`
 	WatchTimeSep int64  `json:"watch_time_sep" value:"10" usage:"Request watch time sep."`
 	PrintBody    bool   `json:"print_body" value:"false" usage:"Print body every request."`
